переделал
ага, добавил попапы, из за чего пришлось на сервер из шареда функцию перекинуть
